Fix printer exit
- Fix action manager exit
- Fix terminating MC port process when not started
- Fix action manager wizard type anotations
- Fix printer exit while in init
- Fix exiting hardware when refresh thread is not running

+++ sl1fw/tests/mocks/mc_port.py
@@ -144,10 +144,11 @@ class MCSerial:
 
         :return: None
         """
-        self.process.terminate()
-        self.process.wait(timeout=3)
-        self.process.stdin.close()
-        self.process.stdout.close()
+        if self.process:
+            self.process.terminate()
+            self.process.wait(timeout=3)
+            self.process.stdin.close()
+            self.process.stdout.close()
 
     def write(self, data: bytes):
         """
